Minor code cleanups and fixing file not found issue
Cleaned up the code to match with my other c++ projects.

Also fixed a bug, where the application exited with a sigsegv, caused by
the application trying to load instructions from a non existing file
and then trying to work with null instructions.

Now the application checks if the file exists first, and only if the
file exists the execution continues.